Add `CityListViewModel`
